requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
datasets==3.3.2
    # via data4vlm (pyproject.toml)
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
filelock==3.17.0
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via
    #   datasets
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
huggingface-hub==0.29.1
    # via
    #   datasets
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
narwhals==1.29.0
    # via altair
numpy==2.2.3
    # via
    #   datasets
    #   pandas
    #   pydeck
    #   streamlit
    #   transformers
packaging==24.2
    # via
    #   altair
    #   datasets
    #   huggingface-hub
    #   streamlit
    #   transformers
pandas==2.2.3
    # via
    #   datasets
    #   streamlit
pillow==11.1.0
    # via streamlit
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via streamlit
pyarrow==19.0.1
    # via
    #   datasets
    #   streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   streamlit
    #   transformers
rich==13.9.4
    # via streamlit
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
safetensors==0.5.3
    # via transformers
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.42.2
    # via data4vlm (pyproject.toml)
tenacity==9.0.0
    # via streamlit
tokenizers==0.21.0
    # via transformers
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.49.0
    # via data4vlm (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   altair
    #   huggingface-hub
    #   referencing
    #   streamlit
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
